Remove accountSigner entirely, and change account initilization options
