+++ src/Account.js
@@ -7,6 +7,16 @@ const computeAccountAddress = require('./computeAccountAddress')
 const encodeFunctionCall = require('./encodeFunctionCall')
 const { ZERO_ADDRESS } = require('./constants')
 const bitmapPointer = require('./utils/bitmapPointer')
+const { signEIP712, constants } = require('@brinkninja/utils')
+const { MAX_UINT256 } = constants
+const {
+  verifyTokenToTokenSwap,
+  verifyEthToTokenSwap,
+  verifyTokenToEthSwap
+} = require('./callVerifiers')
+const {
+  metaDelegateCallSignedParamTypes
+} = require('./constants')
 const BN = ethers.BigNumber.from
 
 const _directCalls = [
@@ -64,22 +74,24 @@ class Account {
   }) {
     this._environment = environment
     this._implementationAddress = _.find(this._environment.deployments, { name: 'account' }).address
-    this._ownerAddress = ownerAddress
     this._accountVersion = this._environment.accountVersion
     this._accountDeploymentSalt = this._environment.accountDeploymentSalt
     this._provider = provider
     this._signer = signer
+    this._chainId = environment.chainId
     this._deployerAddress = _.find(this._environment.deployments, { name: 'singletonFactory' }).address
     this._deployAndExecuteAddress = _.find(this._environment.deployments, { name: 'deployAndExecute' }).address
 
-    this.address = computeAccountAddress(
-      this._deployerAddress,
-      this._implementationAddress,
-      this._ownerAddress,
-      this._accountDeploymentSalt
-    )
-
-    this._accountImpl = this._ethersContract('Account', this._implementationAddress)
+    if (typeof ownerAddress !== `undefined`) {
+      this._ownerAddress = ownerAddress
+      this.address = computeAccountAddress(
+        this._deployerAddress,
+        this._implementationAddress,
+        this._ownerAddress,
+        this._accountDeploymentSalt
+      )
+      this._accountImpl = this._ethersContract('Account', this._implementationAddress)
+    }
 
     this.estimateGas = {}
     this.populateTransaction = {}
@@ -404,6 +416,230 @@ class Account {
 
     return bytecode
   }
+
+  async accountAddress () {
+    const addr = computeAccountAddress(
+      this._findContractAddress('singletonFactory'),
+      this._findContractAddress('account'),
+      await this.signerAddress(),
+      this._environment.accountDeploymentSalt
+    )
+    return addr
+  }
+
+  async signerAddress () {
+    const addr = await this._signer.getAddress()
+    return addr
+  }
+
+  async signCancel(bitmapIndex, bit) {
+    const call = {
+      functionName: 'cancel',
+      paramTypes: [
+        { name: 'bitmapIndex', type: 'uint256' },
+        { name: 'bit', type: 'uint256'}
+      ],
+      params: [bitmapIndex, bit]
+    }
+
+    const signedCall = await this.signMetaDelegateCall(
+      this._findContractAddress('cancelVerifier'), call
+    )
+
+    return signedCall
+  }
+
+  async signEthTransfer(bitmapIndex, bit, recipient, amount, expiryBlock=MAX_UINT256) {
+    const call = {
+      functionName: 'ethTransfer',
+      paramTypes: [
+        { name: 'bitmapIndex', type: 'uint256' },
+        { name: 'bit', type: 'uint256'},
+        { name: 'recipient', type: 'address'},
+        { name: 'amount', type: 'uint256'},
+        { name: 'expiryBlock', type: 'uint256'}
+      ],
+      params: [bitmapIndex, bit, recipient, amount, expiryBlock]
+    }
+
+    const signedCall = await this.signMetaDelegateCall(
+      this._findContractAddress('transferVerifier'), call
+    )
+    return signedCall
+  }
+
+  async signTokenTransfer(bitmapIndex, bit, tokenAddress, recipient, amount, expiryBlock=MAX_UINT256) {
+    const call = {
+      functionName: 'tokenTransfer',
+      paramTypes: [
+        { name: 'bitmapIndex', type: 'uint256' },
+        { name: 'bit', type: 'uint256'},
+        { name: 'token', type: 'address'},
+        { name: 'recipient', type: 'address'},
+        { name: 'amount', type: 'uint256'},
+        { name: 'expiryBlock', type: 'uint256'}
+      ],
+      params: [bitmapIndex, bit, tokenAddress, recipient, amount, expiryBlock]
+    }
+
+    const signedCall = await this.signMetaDelegateCall(
+      this._findContractAddress('transferVerifier'), call
+    )
+    return signedCall
+  }
+
+  async signEthToTokenSwap(bitmapIndex, bit, tokenAddress, ethAmount, tokenAmount, expiryBlock = MAX_UINT256) {
+    verifyEthToTokenSwap(tokenAddress, ethAmount, tokenAmount, expiryBlock)
+    const call = {
+      functionName: 'ethToToken',
+      paramTypes: [
+        { name: 'bitmapIndex', type: 'uint256' },
+        { name: 'bit', type: 'uint256'},
+        { name: 'token', type: 'address'},
+        { name: 'ethAmount', type: 'uint256'},
+        { name: 'tokenAmount', type: 'uint256'},
+        { name: 'expiryBlock', type: 'uint256'},
+        { name: 'to', type: 'address'},
+        { name: 'data', type: 'bytes'}
+      ],
+      params: [
+        bnToStr(bitmapIndex), bnToStr(bit), tokenAddress, bnToStr(ethAmount), bnToStr(tokenAmount),
+        bnToStr(expiryBlock)
+      ]
+    }
+
+    const signedCall = await this.signMetaDelegateCall(
+      this._findContractAddress('limitSwapVerifier'), call
+    )
+    return signedCall
+  }
+
+  async signTokenToEthSwap(bitmapIndex, bit, tokenAddress, tokenAmount, ethAmount, expiryBlock = MAX_UINT256) {
+    verifyTokenToEthSwap(tokenAddress, tokenAddress, ethAmount, expiryBlock)
+    const call = {
+      functionName: 'tokenToEth',
+      paramTypes: [
+        { name: 'bitmapIndex', type: 'uint256' },
+        { name: 'bit', type: 'uint256'},
+        { name: 'token', type: 'address'},
+        { name: 'tokenAmount', type: 'uint256'},
+        { name: 'ethAmount', type: 'uint256'},
+        { name: 'expiryBlock', type: 'uint256'},
+        { name: 'to', type: 'address'},
+        { name: 'data', type: 'bytes'}
+      ],
+      params: [
+        bnToStr(bitmapIndex), bnToStr(bit), tokenAddress, bnToStr(tokenAmount), bnToStr(ethAmount),
+        bnToStr(expiryBlock)
+      ]
+    }
+
+    const signedCall = await this.signMetaDelegateCall(
+      this._findContractAddress('limitSwapVerifier'), call
+    )
+    return signedCall
+  }
+
+  async signTokenToTokenSwap(bitmapIndex, bit, tokenInAddress, tokenOutAddress, tokenInAmount, tokenOutAmount, expiryBlock = MAX_UINT256) {
+    verifyTokenToTokenSwap(tokenInAddress, tokenOutAddress, tokenInAmount, tokenOutAmount, expiryBlock)
+    const call = {
+      functionName: 'tokenToToken',
+      paramTypes: [
+        { name: 'bitmapIndex', type: 'uint256' },
+        { name: 'bit', type: 'uint256'},
+        { name: 'tokenIn', type: 'address'},
+        { name: 'tokenOut', type: 'address'},
+        { name: 'tokenInAmount', type: 'uint256'},
+        { name: 'tokenOutAmount', type: 'uint256'},
+        { name: 'expiryBlock', type: 'uint256'},
+        { name: 'to', type: 'address'},
+        { name: 'data', type: 'bytes'}
+      ],
+      params: [
+        bnToStr(bitmapIndex), bnToStr(bit), tokenInAddress, tokenOutAddress, bnToStr(tokenInAmount),
+        bnToStr(tokenOutAmount), bnToStr(expiryBlock)
+      ]
+    }
+
+    const signedCall = await this.signMetaDelegateCall(
+      this._findContractAddress('limitSwapVerifier'), call
+    )
+    return signedCall
+  }
+
+  async signMetaDelegateCall (toAddress, call) {
+    const signedFnCall = await this.signFunctionCall(
+      'metaDelegateCall',
+      metaDelegateCallSignedParamTypes,
+      [ toAddress, call ]
+    )
+    return signedFnCall
+  }
+
+  async signFunctionCall (functionName, paramTypes, params) {
+    let encodedParams = []
+    for (let i in params) {
+      const typeData = paramTypes[i]
+      if (typeData.calldata) {
+        const callEncoded = encodeFunctionCall(params[i])
+        encodedParams[i] = callEncoded
+      } else {
+        encodedParams[i] = params[i].toString()
+      }
+    }
+
+    const { typedData, typedDataHash, signature } = await signEIP712({
+      signer: this._signer,
+      contractAddress: await this.accountAddress(),
+      contractName: 'BrinkAccount',
+      contractVersion: this._accountVersion,
+      chainId: this._chainId,
+      method: functionName,
+      paramTypes,
+      params: encodedParams
+    })
+
+    const signerAddress = await this._signer.getAddress()
+
+    return {
+      message: typedDataHash,
+      EIP712TypedData: typedData,
+      signature,
+      signer: signerAddress,
+      accountAddress: await this.accountAddress(),
+      functionName,
+      signedParams: parseParams(paramTypes, params)
+    }
+  }
+
+  _findContractAddress (contractName) {
+    return _.find(this._environment.deployments, { name: contractName }).address
+  }
+}
+
+function parseParams (paramTypes, params) {
+  let paramsArray = []
+  for (let i in paramTypes) {
+    const { name, type, calldata } = paramTypes[i]
+    paramsArray[i] = {
+      name,
+      type
+    }
+    if (calldata) {
+      paramsArray[i].value = encodeFunctionCall(params[i])
+      paramsArray[i].callData = {
+        functionName: params[i].functionName,
+        params: parseParams(params[i].paramTypes, params[i].params)
+      }
+    } else {
+      paramsArray[i].value = params[i]
+    }
+  }
+  return paramsArray
+}
+
+function bnToStr (bnOrString) {
+  return bnOrString.toString()
 }
 
 // splits a call for metaDelegateCall into signedData and unsignedData
